+++ packages/bundlekit-plugin-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @bundlekit/plugin-react
 
+## 0.0.10
+
+### Patch Changes
+
+- 0e06ecf: 修复项目的 lib 模式问题
+
 ## 0.0.9
 
 ### Patch Changes